refactor: Cart 리팩토링(#353)
