--- src/main/java/project/trendpick_pro/domain/cart/entity/dto/request/CartRequest.java
@@ -1,9 +1,0 @@
-package project.trendpick_pro.domain.cart.entity.dto.request;
-
-
-import jakarta.validation.constraints.NotBlank;
-
-public class CartRequest {
-    @NotBlank
-    private Long memberId;
-}
